refactor skill button to its own file
