+++ app/src/main/java/io/github/fate_grand_automata/ui/skill_maker/SkillMakerMasterSkills.kt
@@ -1,19 +1,12 @@
 package io.github.fate_grand_automata.ui.skill_maker
 
 import android.content.res.Configuration
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.sizeIn
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -71,33 +64,6 @@ fun SkillMakerMasterSkills(
                     }
                 }
             }
-        }
-    }
-}
-
-@Composable
-fun RowScope.SkillButton(
-    skill: Skill,
-    color: Color,
-    onClick: () -> Unit
-) {
-    Surface(
-        color = color,
-        shape = MaterialTheme.shapes.medium,
-        modifier = Modifier
-            .padding(5.dp)
-            .sizeIn(maxWidth = 45.dp)
-            .aspectRatio(1f)
-            .fillMaxSize()
-            .weight(1f, false),
-        onClick = onClick
-    ) {
-        Box {
-            Text(
-                skill.autoSkillCode.toString(),
-                color = Color.White,
-                modifier = Modifier.align(Alignment.Center)
-            )
         }
     }
 }
